% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{add_months}
\alias{add_months}
\title{Add months to a Date}
\usage{
add_months(x, n)
}
\arguments{
\item{x}{a Date vector}

\item{n}{the number of months that's added to \code{x}}
}
\description{
Add months to a Date
}
