% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-part.R, R/extendr-wrappers.R
\name{date_part}
\alias{date_part}
\alias{year}
\alias{month}
\alias{quarter}
\alias{iso_week}
\alias{mday}
\alias{wday}
\alias{yday}
\title{Extract Date Part}
\usage{
year(ref_date)

month(ref_date)

quarter(ref_date)

iso_week(ref_date)

mday(ref_date)

wday(ref_date)

yday(ref_date)
}
\arguments{
\item{ref_date, }{a Date vector. It will try to convert the input to date via \code{\link[=ymd]{ymd()}},
if the input is not a Date.}
}
\value{
an integer vector
}
\description{
Extract Date Part
}
\section{Detailed explanation}{

\itemize{
\item year, month, quarter: get the year, month, quarter part
\item iso_week: ISO 8601 week starting from 1
\item mday: the day of month starting from 1
\item yday: the day of year starting from 1
\item wday: the day of week (ISO 8601 weekday number, Monday is 1)
}
}

