% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beop.R
\docType{data}
\name{beop}
\alias{beop}
\alias{eop}
\alias{bop}
\title{Begin/End of Period (BOP and EOP)}
\format{
An object of class \code{list} of length 12.

An object of class \code{list} of length 12.
}
\usage{
eop

bop
}
\description{
A list of BOP/EOP functions.
}
\note{
\code{bop} and \code{eop} provides a list of functions. Their names consist of two
letters, the first which means __l__ast, __t__his, __n__ext while the second
means __y__ear, __q__uarter, __m__onth, __w__eek. Thus, \code{eop$ty()} means
"the __e__nd __o__f period of __t__his __y__ear".
}
\keyword{datasets}
