% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beop.R
\docType{data}
\name{beop}
\alias{beop}
\alias{eop}
\alias{bop}
\title{Begin/End of Period (BOP and EOP)}
\format{
An object of class \code{list} of length 12.

An object of class \code{list} of length 12.
}
\usage{
eop

bop
}
\description{
A list of BOP/EOP functions.
}
\note{
\code{bop} and \code{eop} provides a list of functions. Their names consist of two
letters, the first which means \strong{l}ast, \strong{t}his, \strong{n}ext while the second
means \strong{y}ear, \strong{q}uarter, \strong{m}onth, \strong{w}eek. Thus, \code{eop$ty()} means
"the \strong{e}nd \strong{o}f period of \strong{t}his \strong{y}ear".
}
\keyword{datasets}
