% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beop.R
\docType{data}
\name{eop}
\alias{eop}
\alias{bop}
\title{Find the Beginning or Ending of Period}
\description{
Each of \code{bop} and \code{eop} contains a list of functions, whose names all
consist of two letters, the first of which stands for \strong{l}ast, \strong{t}his, \strong{n}ext
while the second stands for \strong{y}ear, \strong{q}uarter, \strong{m}onth, \strong{w}eek.
For example, \code{eop$ty()} means "the \strong{e}nding \strong{o}f \strong{p}eriod of \strong{t}his \strong{y}ear"
and \code{bop$lm()} means "the \strong{b}eginning \strong{o}f \strong{p}eriod of \strong{l}ast \strong{m}onth".
}
\details{
All functions' signatures are the same, with only one argument
\code{x}, which could be a \code{Date} or values that can be converted to \code{Date} via \code{\link[=ymd]{ymd()}}.
}
\examples{
bop$ty(as.Date("2021-03-02"))
## supports 'YMD' formatted integer or string
bop$ty(210302)
eop$tm(200201)

}
\keyword{datasets}
