% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{edate}
\alias{edate}
\title{Calculate the date before / after months}
\usage{
edate(ref_date, months)
}
\arguments{
\item{ref_date}{a Date vector}

\item{months}{the number of months that's added to \code{ref_date}}
}
\description{
Calculate the date before / after months
}
\note{
The function name is the same as the Excel function \code{EDATE()} and
does the same. It returns the date that is the indicated number of months
before or after the ref date.
}
