% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{ymd}
\alias{ymd}
\title{Convert integers or strings to Date}
\usage{
ymd(x)
}
\arguments{
\item{x}{an integerable or string vector in ymd format}
}
\value{
a Date object
}
\description{
Convert integers or strings to Date
}
